Fix plasmoid status when compact plasmoids expand
Sets the status to one that actually exists when the dialog is displayed
(RequiresAttentionStatus).
Always set the status back, sometimes it's actually unknown, and it's still
a valid status to go back to. Otherwise we keep the plasmoid in
RequiresAttentionStatus status and the panel never hides back (if in
auto-hide).

REVIEW: 117724