Bare-bones new Formats KCM
This forms a stab at the the new Locale settings, which need to be
transplanted to QLocale. Currently, this means much coarser setting,
that is no specific settings for the individual items, just a country
setting. From there, we can add back more fine-grained settings,
transplanted from the locale KCM.

Primarily, this makes language settings available to startkde, so the
LC_* variables can be exported.

This is currently non-functional, just a bare-bones KCM, loading a basic
UI file.

Run "kcmshell5 formats" to test.

CCBUG:331930