Use LANG for global settings
According to spec, LANG is used as fallback exactly as we need it. If
detailed is unchecked, we now just set LANG, and be done with that.

We can remove the specialized handling for CTYPE this way as well. We
won't touch it unless the user puts it in her config by hand. In that
case, we export it like everything else.