Reset currentItem in ListViews correctly
Without this patch, all items (apps, actions, etc.) would keep selection
state until someone else "steals" it. The problem is that in many cases,
it's not the mouse position that matters which item gets activated, but
the currently selected item in a given listview.

Adding an onExited handler to the onEntered handler fixes this, and thus
the bug that in some cases, the wrong app gets started.

BUG:332808