read appview's currentItem before animating
moveRight() executed after the animation means we're querying for the
item not after it's clicked, but we have an item start the animation,
and once the animation is done, currentItem is checked, and opened.

Querying listview's currentItem before the animation fixes the wrongly
opened categories.

This should put the final nail in the coffin of
BUG:332808