Simply Kickoff's mouse hover selection
This patch removes the MouseArea behind the listviews, which selected
the current item on hover beyond the listview's bounds. The positioning
in here isn't very reliable, especially when the content scrolls.

Simplify this by moving the margins inside the item and expanding the
listviews to the edges. This keeps the mouse interaction along the edges
intact, gives more expected behavior with the scrollbar, and doesn't
have the subtle positioning problems in finding the item.

This fixes wrongly selected entries in Kickoff when hovering over the
edges of the items, left and right, and fixes scrollbar behaviour.

BUG:332808