Plasma - New Translations KCM
Initial implementation of a new KCM to configure the LANGUAGE envvar
to control translations used by applications under Plasma. The list of
available langauges is currently only those installed for Plasma. We
may later add all installed po files, or allow the user to add a custom
langauge.

Currently this does not create the script for startkde to use to export
the setting.