Formats KCM: Use QLocale::name() when writing to the configuration file
Currently the Formats KCM writes its configuration file using the
selected locale's bcp47Name(). This breaks locale loading as locales are in
the form "foo_FOO" (e.g., "it_IT" for my own locale) and instead the Formats
KCM exports LANG as "foo" (e.g. "it").

This causes a bunch of runtime warnings and locales don't get actually
loaded.

This is just part of the job: more work is needed for encoding and
script handling (see TODO in the comment).

Reviewed-by: Sebastian Kugler, John Layt, Lukas Tinkl

REVIEW: 118584