respect the minimum size
make the applets respect the minimum size, specified by applets, with  the following logic:
if there is a preferred representation, enforce its minimum size,
if there is a compact representation, enforce its minimum size.

note: the minimum size of the full representation is not enforced if it's not the preferred one, because it must still be possible to switch to the compact one.