Remove completley wrong changes to ToolboxButton
Adapt iconsize to toolbox size introduces a blatant binding loop
It was changed to be 0.8 of the frame, but the frame was based on the icon size

The binding loop was then "fixed" by wrapping in an item which doesn't fix the
root problem of it being a loop.

The iconSize comes from units.iconSize.Small in ToolBoxRoot in the first place, making
the entire change pointless. If we do want to scale the icon we just need to change that
one line in ToolBoxRoot.

We certainly don't want to make changes that leave stray unused properties about.

Revert "remove undetected binding loop"

This reverts commit eddb56e5a7cac3cd631f1cb8e315af02f57779ee.

Revert "use standard icons sizes for the toolbox icon"

This reverts commit 216ee387648f7105561dabcf33d06ec880be65a3.

Revert "Adapt icon size to toolbox size"

This reverts commit e38dc4f0dadd5ca40f34e4ba538eedd23c594e92.

BUG: 336441
REVIEW: 119229
CCBUG: 337055