Make Kicker's root item a FocusScope
When opening kicker with a shortcut (e.g. alt-f1) it would not have any
focus. This is because the root item would get the focus but not "pass it
on" to any child below itself.
Making the root item a FocusScope means that when it receives focus the
focus moves to the line edit since that one has focus: true.

REVIEW: 119874