Improve the keyboard layout systray icon
The label-icon used to be blurry and not fitting very well into the
tray; this patch makes it less blury, scaleing with the panel and better
aligned.

Overall it's still not much, ideally this should be a full-fledged
plasmoid as that would allow knowing the exact size of the panel and
thus how big the font needs to be etc. Hardcoding any size is shooting
in the dark. However now the size is "big enough" so it should at least
scale better.

Still, a plasmoid would do much better job.

Reviewed-by: Lukas Tinkl