Make sure that the time and date are editable when ntpdate is missing
This fixes a bug where if you enable automatic time synchronisation and for whatever reason remove ntpdate the time and date are not editable, and you can't disable the automatic synchronisation due to the missing binary. The only way around it currently is to install ntpdate again, go to date and time settings, disable automatic synchronisation and then finally remove ntpdate again.

It also makes the Servers list uneditable when ntpdate is missing, as well as it having the same tooltip as the checkbox.

BUGS: 337707
REVIEW: 120077