Load config before we monitor items for changes
Previous code restored the config from the delegate button which ended
up running after we've set loaded the UI and started monitoring it for
changes.

This resulted in the apply button being always enabled if anything was
different from the default.

This moves the restore into somewhere far more logical, as item
delegates shouldn't be performing logic in some /other/ QML file.

BUG: 338680