Fix Apply button in applet configuration
There's no need to call restoreConfiguration() in Component.onCompleted since the manipulation
of the StackView's sourceFile property in said method will already trigger the onSourceFileChanged
handler which will restoreConfigration() also.

This lead to the issue that for data structures QML cannot compare, like the Array of extra items
in systray config, triggered the change handler on re-assignment by the method running a second
time enabling the Apply button.

REVIEW: 120386
BUG: 339428
FIXED-IN: 5.1