Avoid a lot of syncs to FavoritesModel in kickoff
Current code loaded the list of favourites into a string. It then
readded them one by one syncing the config after each change.
At the end the config is the same as when we started but with 4 syncs in
the meantime.

Splitting add and load methods solves this.

This should hopefully fix lockups as we're no longer writing to
kickoffrc so often on load as well as give us a significant speed boost.

BUG: 339451
REVIEW: 120455