Containment: Redirect wheel events to ContainmentInterface when view is not overflowing.
This makes the wheel scroll Mouse Actions trigger work on Folder.

ScrollView's internal implementation is very greedy for wheel
events (to the point of overriding Flickable completely and making
them unfilterable) irregardless of viewport size, so we hijack
events before they reach it if we know we're not going to need
them there.

BUG:341380