Remove bell kcm
The kcm sets up the xbell, but konsole for example doesn't use the
xbell but knotify. The code is also xcb only, I have no idea if it could
be ported to wayland. For users the knotify way of treating the bell
(visual/sound/custom executable) should satisfy all needs.

Then there is also the accessibility kcm allowing to switch on and off
the system bell (using xbell as well) and allowing to flash the screen
in a color (I could not get that one to work, neither using xterm nor konsole).

REVIEW: 122464