When setting colour scheme information for KDE4, don't read from KF5 kdeglobals
When krdb sets the colour scheme for KDE4 applications, it should not
use the KF5 global or system configuration, as it won't be in the
correct format.  Instead, setup KConfig instances to only read the
file given.

This also works around bug 340691 in KConfig, which fixes bug 343583.
Note this change should be left regardless of when KConfig is fixed,
as KF5 settings remain unneeded in KDE4 configuration files.

BUG: 343583
FIXED-IN: 5.2.1