Add screen reader to Accessibility KCM
This patch adds a simple check box to the KCM which enables the screen
reader.
It additionally sets up the default shortcut - Meta-Alt-S which is the
same as in Gnome. Using the same shortcut is of course extremely
important, since blind users rely on an easy way to activeate the screen
reader.

For now the whole process is hard-coded to Orca since there are no real
alternatives for Linux screen readers at the moment.

The gconfig call which disables the screen reader will make Orca exit.
It still needs to be launched manually.

REVIEW: 122540