Add systemd support into the clock KCM as an optional dependency
The current time setting helper is incredibly broken.

It manually tries to run a range of NTP utilities, all of which are
deprecated.

We can just call timedated directly and cut out the middleman as it has
uses polkit anyway.

This is currently an optional dependency, and the original helper still
exists. It makes the code messy, but we have users to support for now.

Detection is done at runtime

BUG: 196316
BUG: 311286
BUG: 317784
BUG: 319072
BUG: 337012
BUG: 339582
BUG: 241817
BUG: 178968
BUG: 320456
BUG: 317999
FEATURE: 337659
REVIEW: 122400