Adding libkactivities-stats to the build
The experimental library can not land in the kactivities framework just
yet. Since it is to be used by plasma-desktop in the next release, we
need to include it (for the time being) in p-d.

The idea came from Eike. While it has downsides, it also has a few
benefits: - no need to have a monthly release cycle, so we can actually
do work without much bureaucracy surrounding API and ABI compatibility;
- when the library becomes stable, it will move to KActivities and lose
the 'experimental' part of the name. This means that the distributions
will not have issues of two packages providing the same files; - no need
to sync the releases and think about which KActivities version plasma
5.x will be used against; - I guess I'm forgetting more benefits. :)

The patch includes update.sh script that syncs this code with that in
KActivities. (the development should still be in the KActivities
ivan/libkactivities-experimental-stats branch)

libkactivities-stats needs to be tested before going into the main
repository. Until that happens, it lives in a branch of kactivities
that is synced here.

REVIEW: 123088