Remove pointless timer
There is a connection on backgroundHints changing, we don't need to put
the initial setting on a timer, it will update anyway. The comment next
to it made no sense, we know the plasmoid top level component is
comlete.

Testing shows everything is faster as we don't paint a background on
notes only to remove it again.