Improve Applet Alternatives dialog
For some reason setting the currentIndex does not work properly. I tried both in the delegate
and the view but neither worked. So, a Timer comes as a last resort.
Also implement keyboard navigation, you can now use the arrow keys to navigate through the
list and Return to accept or Escape to cancel. To achieve this we rely on the currentIndex
rather than the checkable property.

FIXED-IN: 5.3.2
BUG: 345786