Fix Phonon Pulseaudio speaker test button sound
During the 5.x port a signal-slot connection became a slot-slot connection
as both the signal and the slot used the same name which prevents the
test sound from working as the signal is never connected to anything.

Renamed the affected slot to resolve ambiguity so now the widget is back to
a signal-slot connection.

BUG: 349825
FIXED-IN: 5.3.95