[Panel] Fix min/max size sliderhandles
The x (or y) position to value calculation differed from the
value to x (or y) calulation.
This lead to strange behavior if the sliders were moved slowly (only
when panel alignment is center): Though the mouse was moved slowly,
the sliders moved very fast, sometimes even in the opposide direction.

REVIEW: 124481