[Panelconfig] Reset the panel offset on alignment change
As offset and length have a different meaning in all alignments, the
panel shifts on alignment change. This could result in wrong panel
positions (e.g. panel shifted over a monitor border). A better approach
would be a recalculation of all values, so that the panel stays at it's
current position, but this would be error prone and complicated. As the
panel alignment is rarely changed, it's not worth it. The more easy
approach is just setting the panel offset to zero. This makes sure the
panel has a valid position and size.

REVIEW: 124785