[Panelconfig] Add more / fix constraints in the position/size handles
Prevent users from moving the panel off the monitor, and prevent users
from moving a panel with center alignment to the left (right) and then
pull one of the right (left) sliders to the right (left). Because the
left and right sliders are coupled, this would make the left (right)
sliders to go off the monitor.

Also fix that the left (right) sliders cannot be moved over the middle.
(In fact they can by using the offset slider, but as soon as on of the
left (right) handles are dragged, it sets the position back to the
middle)

Also delete the default values for min/maxPosition in the SliderHandle,
as every handle sets its own min/maxPositions now.

REVIEW: 124786