Use PlasmaIcon so that it can do the colour themeing
In order to let plasma do the icon colour changing we need Plasma to get
the icons as strings, so this refactors the model.

Contact avatars require pictures so abstract method stores icon as a
variant.

This should be a tiny bit faster too as QIcon::fromTheme does some
lookups to see if it's valid, whereas strings will be loaded on demand.

DO NOT MERGE: Whilst PlasmaIcon does the colour changing, it sucks as
1) there's no fallback icon if we can't find something
2) the animations look silly in this context
3) use of QIconItem might be deliberate?