Fix loading/saving in kded kcm
Fix loading/saving of the autoload status of kded startup services in
the kcm.

Loading was broken due to a KPluginMetaData::pluginId() being passed to
autoloadEnabled(), instead of a KPluginMetaData.

Saving was broken due to comparing KPluginMetaData::fileName() to
treeitem->data(LibraryRole). This didn't work due to
KPluginMetaData::fileName() sometimes containing (part of) the path in
addition to the plugin/library name. This was solved by using
KPluginMetaData::pluginId() instead of KPluginMetaData::fileName().

BUG: 346848
BUG: 347429
REVIEW: 124840