Fix widget explorer sometimes opening on the wrong side of the screen.
Work around a Qt bug.

If there is no setGeometry call on QWindow the XCB backend will not pass
our requested position to kwin
as our window position tends to be 0, setX,setY no-ops and means we
never set a position as far as QWindow is concerned
by setting it to something silly and setting it back before we show the
window we avoid that bug.

Qt patch also filed (https://codereview.qt-project.org/123989)
so hopefully this can be removed in future if that gets merged.

BUG: 332702
REVIEW: 124869