Fix KF5 port regression in kcm_standard_actions
The KShortcutsEditor reads the property "defaultShortcuts" from the
assigned QAction. This commit changes KCM code to correctly fill in
QActions with default "hardcoded" shortcuts.
Previously it was appending the defaults to the list of active
shortcuts, but that just creates shortcut lists that grow longer and
longer every time user saves settings in this KCM.

BUG: 343665
FIXED-IN: 5.4.1
REVIEW: 123828