kcm_keyboard: Fix behavior of layout config edit delegates
By default, commiting edited value in QStyledItemDelegate only occurs
when leaving the editor, eg. on pressing Enter or moving to other item.
This makes the delegates commit updated value to model as soon as value
in editor is changed.

This also fixes the size of label editor (it now fills the entire cell)
and refresh the label icons when refreshing model.

BUG: 350757
FIXED-IN: 5.4.2
REVIEW: 125405