kcm_formats: optimize loading of flag icons
Keep a local cache of the loaded icons, so there is only one
QStandardPaths lookup and one QIcon loading for each flag code.
This has the nice advantage to reduce the I/O operations a lot (making
the startup of the module noticeably faster), and to reduce the memory
usage (using the implicit sharing of QIcon).

The "unknown" icon is cached as well.