[Widget Explorer] Make it feel a lot quicker
Widget Explorer takes quite a while to open, especially now that we initially load double
the number of items (two columns) and significantly more complex SVGs.

This patch uses a couple of tricks to make it feel a lot faster:
- Create the context menu items on demand
- Set the model for the list after a slight delay
- Some minor cleanups

By setting the model later we let the dialog slide in first (KWin does the animation
anyway so it won't block) and then nicely fade in the view (with an Animator) once we
have loaded it.

CHANGELOG: Widget Explorer opens quicker now

REVIEW: 126163