kacess should not be built if xkb is not present
The XCB-XKB dependency of plasma-desktop is optional, however kacess
depends directly and unconditionally on the xcb/xkb.h header file.
This causes the build of plasma-desktop to fail at compile stage if the
xkb header is not present.

As per the convention in the kcm modules shipped by plasma-desktop,
inclusion of the kacess subdirectory is made conditional on XKB being
present and properly detected during the cmake configure stage.

Patch by: Johan Ouwerkerk
REVIEW:126013