[Panel Containment] Hide applet if its status is Hidden
With the newly introduced HiddenStatus we now have a way to hide panel applets in a
controlled way instead of incorporating hacks like the Pager currently does.

When an applet sets its status to Hidden it will be removed from the layout and
fully collapse as if it wasn't there. However, if you open the panel configuration,
it shows up so you can still configure or permanently remove the applet.

REVIEW: 126776