[Task Manager] Provide media controls in tooltips
This adds media player controls to the tooltip of a media player, using the already
existing mpris dataengine used by the Media Controller applet.

When album art is available, no window thumbnail will be shown, instead the album art.
There will be a close buton on the album art. If the player window is hidden but the player
supports the MPRIS "Raise" command and it is pinned to the task manager, its window can
still be brought to the front, alleviating the need for a system tray icon.

REVIEW: 126647
BUG: 352126
FIXED-IN: 5.6.0