[Widget Explorer] Show desktop on dragging widget
When starting to drag a widget, show desktop mode is enabled causing all windows
but the widget explorer to disappear.

Windows are only restored when widget explorer closes, not as soon as you drop it,
so you can see where you dropped the widget and configure it right away.

REVIEW: 126938
BUG: 327530
FIXED-IN: 5.6.0