[kcms] Adjust keyboard kcm so that it works on Wayland
Summary:
So far the keyboard kcm was forced to X11 only. But that is not required,
also on Wayland xkb is used for keyboard layouts. Thus everything which
is not X11 specific can also work on Wayland. In fact KWin reads the
keyboard layout information from the configuration file this kcm writes
to.

The important change here is that a default fallback rule name (evdev)
is used. On Wayland we are not yet able to resolve the rule name thus
the complete UI was disabled. With this change the UI is configurable
and works - one can change the layout in KWin.

Of course in the long run this needs better Wayland integration to
properly resolve the rule name.

Reviewers: #plasma

Subscribers: plasma-devel

Projects: #plasma

Differential Revision: https://phabricator.kde.org/D1006