[kcms] Make cursortheme work on Wayland
Summary:
Loading and displaying the cursor images is not X11 specific.
Also the stored configuration is used in a Wayland session, thus the
kcm is useful and should work.

The adjustments make sure that all themes are listed and we don't hit
an unconditional X11 specific branch.

Currently not working on Wayland is:
* live preview of icons
* updating running applications (not supported by QtWayland)

KWin and newly started applications pick up the changed cursor theme.
As not all applications are adjusted, the information for session restart
is shown.

Reviewers: #plasma

Subscribers: plasma-devel

Projects: #plasma

Differential Revision: https://phabricator.kde.org/D1014