[Applet / Wallpaper Configuration] Load config page with initial cfg properties already set
Rather than creating the config page and then setting the properties afterwards have StackView
load the item with all the cfg properties already set.

This prevents situations where a binding might limit a value to some range based on some config
which isn't set by the time the limit is enforced as can be seen in the slideshow interval config.

Moreover, it makes Component.onCompleted in the config pages actually useful, I hope it will
make dealing with ComboBoxes less tedious. On the downside it might break applet configs that
rely on cfg_fooChanged to populate themselves which previously was required as at the time of
Component.onCompleted the cfg_ properties were not yet populated.

Differential Revision: https://phabricator.kde.org/D1220