Workaround searching lockup
Summary:
QQC::ScrollView has an annoying bug where you can get it in a situation
where if you have exactly if delegates are based on the viewport width you
can get in a loop where you need the scrollbar, that shrinks the view so it
resizes the delegates at which point you don't need the scrollbar, so it resizes the
delegates so and so on.

Keeping scrollbars always on is a good workaround.
If anything it's nice to not have things jumping around slightly when searching

BUG: 365868

Test Plan:
Plasma explorer still works.
Approach works as I've used it in the wallpaper selection dialog
for exactly the same reason.

Reviewers: #plasma

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D2239