[Kickoff] Fix start row for drag not always being correct
Summary: The start row for the drag is set to the current index but the current index is not always the item being dragged. This is because syncTarget has not been called yet to update the current index to the item under the cursor. Fix this by calling syncTarget before reading the current index.

Test Plan: Drag items around and notice that the item being moved is not always correct.

Reviewers: broulik, #plasma

Reviewed By: broulik, #plasma

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D2322