User interface for adding launchers as global shortcuts
Summary:
add a menu to pick applications from the start menu structure to
add launchers to be usable as shortcuts.
If an application has jumplist options they will be available as actions.

still need a dbus action to tell the daemon to reload
and probably launchers vs runtime actions will have to be
separed in the ui as well, that may influence the daemon api

Reviewers: graesslin

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D2111