[Kickoff] Fix being unable to reorder entries in favorites menu after scrolling down
Summary: BUG: 345741

Test Plan:
I have many entries in the favorites menu. I can reorder the entries if the vertical scrollbar is at the top but if I scroll to the bottom of the favorites menu and try to reorder the items, it doesn't show a drop indicator for anywhere other than the end of the favorites list.

Steps to Reproduce:
1. Add lots of applications to favorites such that less than half of the favorites are shown when you scroll to the top of the favorites menu list
2. Scroll to the bottom of the favorites list
3. Click and drag the 3rd-last item up one position

Reviewers: bshah, #plasma, broulik

Reviewed By: #plasma, broulik

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D2337