[Applet Configuration] Match category sidebar style with Breeze widget default style
This adjusts the category sidebar to look like the default Breeze widget style.
The width of the sidebar also adjusts (up to a certain maximum width) based on
the labels within so it doesn't grow as wide anymore.
Moreover, keyboard navigation is greatly improved by indicating focus on the category
sidebar and changing categories with arrow keys. Also, when tabbing through the dialog,
the view is scrolled, if the item with focus would otherwise be inaccessible.

As Plasma doesn't seem to be supporting Qt high-dpi scaling anytime soon this uses
units to get the sizing identical to what it looks like in the other applications.

Differential Revision: https://phabricator.kde.org/D2398