add a preview button and a fullscreen preview
Summary:
preview button for themes, launches a fullscreen preview for the theme
the preview image is shown as big as possible, either 1:1 or
scaled down to screen size.

clicking anywhere (or in the top right close button) dismissed the
fullscreen dialog

the thumbnail uses a plasma toolbutton, but in that particular point is fine
to use a plasma control, because it's on top of an image, it's
a custom graphics zone, where on the other hand a system style
toolbutton wouldn't work well.

Reviewers: #plasma, davidedmundson

Reviewed By: #plasma, davidedmundson

Subscribers: davidedmundson, plasma-devel

Tags: #plasma, #vdg

Differential Revision: https://phabricator.kde.org/D2550