[Applet Configuration] Only scroll down if item is barely visible
The logic for scrolling a focus item into view was mostly meant for keyboard-navigating through
a list of CheckBoxes. However, an inner ScrollView - like plasma-pa has in its config dialog - can
also gain activeFocus. This had us try to scroll the ScrollView into view, usually scrolling away
other view contents.

With the added logic we only try to scroll if the control is barely visible at the bottom but
not if significant portions of it are already visible.

Differential Revision: https://phabricator.kde.org/D2585