Add an option to limit the Pager to the current screen.
Summary:
The Pager's display normally encompasses all connected screens
and shows windows on all of them. This adds an option that
restricts display to the plasmoid.screenGeometry rectangle and
enables window filtering for it in the backend.

By user request.

Completely untested as I only have one monitor.

BUG:356693

Reviewers: #plasma, davidedmundson, graesslin

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D2720