xcursor discovery: modernize
Summary:
First we remove the fallback code for a version of Xcursor that
corresponds with an X server KDE doesn't even support any more.

Then we reverse the order of tilde expansion and deduplication, which
before was essentially incorrect.

Next we do deduplication properly, using Qt's removeDuplicates function.

Finally we replace the ugly regex with the helper function in
kcoreextras via std::transform.

This work here reflects the recent review and discussion that went into
the nearly identical code in kde-gtk-kcm.

Test Plan: Observe that cursors are correctly populated.

Reviewers: mart

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D2772