use KPluginSelector to enable runners
Summary:
the configuration dialog for enabling runners by category
was a neat idea but had several problems, gave by it
reinventing the wheel too much.
when changing the locale, the enabled categories list was resetted
due to categories being localized. it was also pretty
difficult to disable a particular plugin as you would have to
know what categoriesyou have to disable.

this approach is more basic, but reusing a common implementation,
we avoid some bugs and misbehaviors.

BUG:350779

Test Plan: ran on an existing krunnerrc with different sets of categories enabled, tested krunner and its kcm with it, all expected runner plugins were used both in the kcm and while searching with krunner

Reviewers: broulik, #plasma, davidedmundson

Reviewed By: #plasma, davidedmundson

Subscribers: davidedmundson, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D2872