Don't create attached LayoutMirroring on non Item
Summary:
The current code tries checking if we're in RTL on the current object,
but is accidentally reading an attached property on the Translate object
rather than the root Item. This gives a warning, and more importantly
won't work.

BUG: 369836

Test Plan:
Ran plasmsa and opened applet handle
Ran plasmashell --reverse and opened applet handle

Reviewers: #plasma

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D2927