Add an option to force alphabetic sorting to Kicker/Kickoff.
Summary:
And default to on.

I don't like this, but downstream complains that their users get
confused when using kmenuedit to edit their menu and newly-
installed apps are no longer sorted-in alphabetically in the
presence of the kmenuedit-generated files causing sorting to be
disabled. And we don't have the resources to fix kmenuedit.

Reviewers: mart

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D2940