Remove line that does nothing
Summary:
AppletInterface is designed to expose Applet methods to QML space.

Emitting a signal on appletinterface from QML doesn't acheieve anything
as all we'll do is notify the QML code which we're already in.

We intention was probably to re-evaluate anything connected to
Applet::contextualActionsAboutToShow, but that's not happening.

Test Plan: Opened toolbox, it looked the same

Reviewers: #plasma

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D3165