Close popups on escape
Summary:
A lot of applets have independent handling making escape close the
expanded representation. However not all of them do.

This leads to both duplicated code and inconsistencies, so we should
tackle this at the source.

If an applet was already using escape internally it will be accepting
the event, and so this code will not cause any regressions.

CCBUG: 362657

Test Plan: Opened a bunch of popups. Pressed escape

Reviewers: #plasma

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D3256