Fix 2 RTL bugs
Summary:
Bug 364778: (DashboardRepresentation.qml)
I add LayoutMirroring.

Bug 372717:
The kicker (Application menu) had special configuration - when the panel is on the right edge the layout is mirrored, because else the sidebar blocking the sub menus.
In RTL direction system this fix cause the search bar to overlap the sidebar.
So I Fix that and I add the same configuration for RTL direction system with left edge panel.

Test Plan:
I compiled from source and use it
For the Application menu I check In RTL and LTR directions and check on all the edge and try to search something
All work fine.

Reviewers: #plasma_workspaces, mart

Reviewed By: mart

Subscribers: hein, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D3553