Make AppstreamQt optional
Summary:
Normally I would just close any bug report saying a 3rd party lib should
be optional, but it's quite a big dependency chain for a fairly minor
feature that is easy to ifdef out.

Also we're depending on something that was only released a few weeks ago,
in experience that generally will cause some distro to patch it out anyway.

BUG: 374310

Test Plan:
My AppstreamQt (from distro packages) is broken, so needed to do a temporary workaround anyway.
Built this and it works, haven't been able to test a with appstream version.

Reviewers: #plasma, apol

Reviewed By: apol

Subscribers: anthonyfieroni, huber, hein, mak, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D3923