[Folder View] Fix right click erroneously opening files
It turns out mouse.buttons is 0 for "clicked" events whereas mouse.button is properly set.
It should bail out when released with right mouse button as that opens the context menu
but this condition is never met.

According to documentation this looks intentional, at least it says "For mouse move
events […]. For mouse press and double click events […]. For mouse release events […]."
I can see no mention of regular "clicked", no idea why that is, though.

BUG: 360219
FIXED-IN: 5.8.6