Restore Menubar style config in application style
This restores the configuration for whether menus should be shown in the application (no
global menu), as title bar button, or as "widget".

The notification popup is removed and replaced by a simple KMessageWidget.
The option "Application Menu widget" doesn't actually do anything, it just enables global
menu but not visible to the user. Perhaps it should show a hint, like "Place the Application
Menu widget in a panel" or something like that.

Differential Revision: https://phabricator.kde.org/D3216