Show descriptive text when changing containment plugins
Summary:
When one changes containment plugin, the rest of the configuration
options don't apply or in some cases don't work - changing a wallpaper
and containment type will result in the old containment getting updated
only to be immediately destroyed.

Fixing that is quite difficult (I spent a while, but not got anything
tidy).

So instead make an intermediate solution of at least not presenting the
user with broken options.

CCBUG: 355642

Test Plan: Ran it

Reviewers: #plasma, mart

Reviewed By: mart

Subscribers: broulik, mart, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D4034